refactor(instrumentation): fold TransitionTracker into the observers file
