+++ src/workspace/workspaceStateLogger.ts
@@ -1,4 +1,4 @@
-import { TransitionTracker } from "../util/transitionTracker";
+import { TransitionTracker } from "../instrumentation/workspace";
 
 import type {
 	Workspace,
